Adding agenda item
Here's a use case where the TSC could potentially be helpful? At least I'd like to understand if we could be:

Some NASA folks had a paper in at SciPy and apparently some of their NASA employers or collaborators had retro-actively a copyright issue with this.
What astonished me is the remark that according to the author at the end of this issue page claims that things got more restricted in recent years. This clashes for me with the noise I'm hearing from NASA that everything goes open source with easy licenses like MIT or similar?
I'd like to discuss this a bit during the meeting.