I couldn't wait, so I started enacting the changes we talked about.
This merges some material from the README into the TSC-Charter, so
requires a 'majority' vote before it can be merged.  Since we are
currently 7 members, in addition to this PR (my vote), it requires
3 approvals before it can be merged (instead of the usual 2).

I moved the Mission statement to the top of the README, eliminating the
need for a separate file.  Likewise, our current 'Code-of_Conduct'
seems to fully serve the purpose of the 'Moderation-Policy' document,
so I removed it as redundant.

This should set the stage for further refinement of TLP and TLWG
documents.