This work restructures a lot of our documents to be aligned with our evolving concepts of TLPs and TLWGs, and to provide clearer information for TLP and TLWG applications and workflow.  We had information spread all over our files, some duplicated, some scattered.  This change consolidates information and policies so that they only appear in one place and are referenced from others.
This change restructures the contents of Bootstrap-Policies so that they are hopefully easier to understand and then copy and use to jump-start TLP and TLWG documentation.  The files there were genericized to apply to both TLPs and TLWGs.  Material in the 'Governance' file was merged into the Charter document, and the 'Moderation-Policy' was merged into the 'Code of Conduct' which really just points to the TSC Code of Conduct.

Combined information from other places into the Application Template, so that it is hopefully more useful.

Reworked the Lifecycle document, and invented a process for TLP/TLWG approval.

Created a proper Contributing document for the TSC.